update navigation of tokensales

+++ app/src/main/java/network/o3/o3wallet/API/O3/O3APIDataModels.kt
@@ -63,7 +63,7 @@ data class TokenSale(val name: String,
                      val footer: Array<FooterRow>)
 
 data class AcceptingAsset(val asset: String,
-                          val basicRate: Long,
+                          val basicRate: Double,
                           val min: Double,
                           val max: Double)
